+++ components/ArticlePageLayout.js
@@ -15,13 +15,13 @@ import useCurrentUser from 'lib/useCurrentUser';
 import * as FILTERS from 'constants/articleFilters';
 import ArticleItem from 'components/ArticleItem';
 import FeedDisplay from 'components/FeedDisplay';
-import Filters from 'components/ListPage/Filters';
-import ArticleStatusFilter from 'components/ListPage/ArticleStatusFilter';
-import CategoryFilter from 'components/ListPage/CategoryFilter';
-import ReplyTypeFilter from 'components/ListPage/ReplyTypeFilter';
-import TimeRange from 'components/ListPage/TimeRange';
-import SortInput from 'components/ListPage/SortInput';
-import LoadMore from 'components/ListPage/LoadMore';
+import Filters from 'components/ListPageControls/Filters';
+import ArticleStatusFilter from 'components/ListPageControls/ArticleStatusFilter';
+import CategoryFilter from 'components/ListPageControls/CategoryFilter';
+import ReplyTypeFilter from 'components/ListPageControls/ReplyTypeFilter';
+import TimeRange from 'components/ListPageControls/TimeRange';
+import SortInput from 'components/ListPageControls/SortInput';
+import LoadMore from 'components/ListPageControls/LoadMore';
 
 const DEFAULT_REPLY_REQUEST_COUNT = 1;
 const MAX_KEYWORD_LENGTH = 100;
